refactor: add types for controller.ts
